requirements.txt
# AWS SDK
boto3==1.34.19
botocore==1.34.19

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AWS Specific
aws-lambda-powertools==2.32.0
mangum==0.17.0

# Data Processing
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
pytest-mock==3.12.0
moto[all]==5.0.0
freezegun==1.4.0
faker==22.0.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0

# Logging
python-json-logger==2.0.7

# CSV Export
pandas==2.2.0
